requirements.txt
aiogram==2.25.1
environs==8.0.0
redis==5.0.4
asgiref==3.8.1
psycopg2-binary==2.9.9
aiohttp
aiosignal==1.3.1
amqp==5.2.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
billiard==3.6.4.0
black==24.4.0
blinker==1.8.1
Brotli==1.1.0
bs4==0.0.2
cachetools
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cron-descriptor==1.4.3
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.3
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.2.1
greenlet==3.0.3
identify==2.5.35
idna==3.6
importlib_metadata==7.1.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.6
locust==2.26.0
lxml==5.2.1
MarkupPy==1.14
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
nplusone==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
polib==1.2.0
pre-commit==3.7.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.22
pyproject-api==1.6.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
tornado==6.1
tox==4.14.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.2
xlrd==2.0.1
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
drf-yasg==1.21.7
django-modeltranslation==0.18.11
djangorestframework==3.14.0
django-filter==23.5
django-rosetta==0.10.0
sentry-sdk==1.40.5
celery==5.2.7
django-celery-beat==2.6.0
beautifulsoup4==4.12.3
django-import-export==3.3.8
django-solo==2.2.0
pillow==10.3.0
django-recaptcha==3.0.0
docxtpl==0.16.8

# telegram bot
python-telegram-bot==13.0

# use only 6.6.1 version, others raise warning
django-ckeditor==6.6.1
django-debug-toolbar==4.3.0
django-querycount==0.8.3
django-prometheus==2.3.1
django-auditlog==3.0.0
django-rosetta==0.10.0
gunicorn
